# Optimize normalization caching and copying