@aabmass 2024-09-11
> Since our exporter is marked non-mutating, we can't change the data point that is passed to us

Would the code be significantly simpler if we could mutate? And are there any negative consequences of marking the exporter as mutating?