@dashpole 2024-09-12
> Would the code be significantly simpler if we could mutate? And are there any negative consequences of marking the exporter as mutating?

That might be a good idea to explore as well.  It looks like it only matters if there are multiple exporters in the config: https://github.com/open-telemetry/opentelemetry-collector/blob/2166b11137ec473b01b6cc5a0b9bf932cb752cf2/internal/fanoutconsumer/metrics.go#L21.  If there is only a single one, there is no additional clone involved.

We already are doing a full clone for GMP anyways: https://github.com/GoogleCloudPlatform/opentelemetry-operations-go/blob/a6eab1eadf29840b84f651dfca699765aa086290/exporter/collector/metrics.go#L342.  Given that, it seems like it would be better to just mark our exporter as mutating, and simplify those things.

> Could we wrap the pdata struct to hide this complexity to the caller?

We probably could, but that is a lot of boilerplate.