@dashpole 2024-09-12
Lets try https://github.com/GoogleCloudPlatform/opentelemetry-operations-go/pull/892 first, and then see how much more of a benefit we get from some of the optimizations here.